aws_cv_clip/requirements_visualization.txt
matplotlib>=3.5.0
Pillow>=9.0.0
numpy>=1.21.0
